+++ Sources/Editor/EditorInspectorView.swift
@@ -6,7 +6,7 @@ struct EditorInspectorView: View {
 
     var body: some View {
         VStack(spacing: 0) {
-            ScrollView {
+            ScrollView(showsIndicators: false) {
                 VStack(alignment: .leading, spacing: 0) {
                     // MARK: Tools
                     VStack(alignment: .leading, spacing: 10) {